LInDT/requirements.txt
cuda11.0
cudatoolkit==11.0.221
cudnn==8.0.5
python==3.8.6
pytorch==1.7.1
dgl-cuda11.0==0.5.3
numpy==1.19.2
scipy==1.5.2
scikit-learn==0.23.2
numba==0.52.0